+++ libs/cmcd/src/CmcdReporter.ts
@@ -1,4 +1,4 @@
-import type { Request } from '@svta/cml-utils'
+import type { HttpRequest } from '@svta/cml-utils'
 import { uuid } from '@svta/cml-utils'
 import { CMCD_DEFAULT_TIME_INTERVAL } from './CMCD_DEFAULT_TIME_INTERVAL.ts'
 import { CMCD_PARAM } from './CMCD_PARAM.ts'
@@ -42,7 +42,7 @@ function createEncodingOptions(reportingMode: CmcdReportingMode, config: CmcdRep
 	}
 }
 
-function defaultRequester(request: Request): Promise<{ status: number; }> {
+function defaultRequester(request: HttpRequest): Promise<{ status: number; }> {
 	const { url, ...init } = request
 	return fetch(url, init)
 }
@@ -106,7 +106,7 @@ export class CmcdReporter {
 	}
 
 	// TODO: Should this be an event handler?
-	private requester: (request: Request) => Promise<{ status: number; }>
+	private requester: (request: HttpRequest) => Promise<{ status: number; }>
 
 	/**
 	 * Creates a new CMCD reporter.
@@ -116,7 +116,7 @@ export class CmcdReporter {
 	 *                    The default is a simple wrapper around the
 	 *                    native `fetch` API.
 	 */
-	constructor(config: Partial<CmcdReporterConfig>, requester: (request: Request) => Promise<{ status: number; }> = defaultRequester) {
+	constructor(config: Partial<CmcdReporterConfig>, requester: (request: HttpRequest) => Promise<{ status: number; }> = defaultRequester) {
 		this.config = createCmcdReporterConfig(config)
 		this.data = {
 			cid: this.config.cid,
@@ -239,7 +239,7 @@ export class CmcdReporter {
 	 * @param req - The request to apply the CMCD request report to.
 	 * @returns The request with the CMCD request report applied.
 	 */
-	applyRequestReport(req: Request): Request {
+	applyRequestReport(req: HttpRequest): HttpRequest {
 		if (!req || !req.url || !this.config.enabledKeys?.length) {
 			return req
 		}