+++ libs/cmcd/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Changed
 
+- Update `CmcdReporter` to use `HttpRequest` from `@svta/cml-utils` (renamed from `Request`)
 - Refactored `Cmcd` type to be the intersection `CmcdRequest & CmcdResponse & CmcdEvent`
 - Restored `CmcdRequest`, `CmcdResponse`, and `CmcdEvent` as separate, standalone types
 - `CmcdRequest` contains the common, object, request, session, and status keys